se modulo el componente timer, y los elementos html
