+++ elements.js
@@ -1,0 +1,5 @@
+export const moves = document.getElementById("moves-count");
+export const timeValue = document.getElementById("time");
+export const buttonControl = document.querySelector(".button-control");
+export const controls = document.querySelector(".game-stats");
+export const gameContainer = document.querySelector(".game-board");
